refacto(FRONT): Rebuild and continue Acount
