+++ Client/src/packages/feat-Account/API.d/user_info.ts
@@ -1,0 +1,22 @@
+
+const getUserImageProfile = () => {
+    return ''
+}
+const getFullname = () => {
+    return 'mohammed es-sadki'
+}
+const getUsername = () => {
+    return 'mes-sadk'
+}
+  
+type API = {
+    getUserImageProfile: () => string
+    getUsername: () => string
+    getFullname: () => string
+}
+  
+export const api: API = {
+    getUserImageProfile,
+    getUsername,
+    getFullname
+}
