+++ Client/src/packages/feat-Account/components/OwnerProfile.tsx
@@ -1,0 +1,6 @@
+import { Avatar, FormControlLabel, Switch, TextField } from "@mui/material";
+
+const OwnerProfile = () => {
+  return <></>
+}
+export default OwnerProfile
