+++ Client/src/packages/feat-Account/components/index.ts
@@ -1,1 +1,4 @@
-export * from "./utils";
+import style from './styled-components.d/Profile.sx'
+
+export * from './styled-components.d'
+export default style
