+++ Client/src/packages/feat-Account/components/styled-components.d/index.ts
@@ -1,0 +1,1 @@
+export * from './Profile.sx'
