+++ Client/src/packages/feat-Account/images_uploads/image-env.ts
@@ -1,0 +1,3 @@
+
+export const Logo = "./federation.png"
+export const Pic = "./Picture.png"
