+++ Client/src/packages/feat-Account/images_uploads/index.ts
@@ -1,0 +1,2 @@
+import Logo from "./federation.png";
+import Picture from "./Picture.png";
