+++ Client/src/packages/feat-Account/index.ts
@@ -1,0 +1,2 @@
+export * from "./Profile"
+export * from "./Settings"
