--- Client/src/packages/feat-Account/index.tsx
@@ -1,3 +1,0 @@
-export * from "./Settings";
-export * from "./Profile";
-export * from "./Settings";
