+++ Client/src/packages/feat-Account/styles/general-Acount.css
@@ -1,0 +1,6 @@
+/* ========================= */
+/*  GENERAL ACOUNT STYLING */
+/* ======================= */
+
+* { }
+
