+++ Client/src/packages/feat-Account/styles/styled-components.d
@@ -1,0 +1,1 @@
+/home/sdk-meb/Desktop/ft_transcendence/Client/src/packages/feat-Account/components/styled-components.d
