+++ backend/.env
@@ -1,7 +1,7 @@
 # run locally Database
-DATABASE_URL="postgresql://user:password@localhost:5432/ft_tran?schema=public" 
+#DATABASE_URL="postgresql://user:password@localhost:5432/ft_tran?schema=public" 
 # run Database with docker
-# DATABASE_URL="postgresql://user:password@postgres:5432/ft_tran?schema=public" 
+DATABASE_URL="postgresql://user:password@postgres:5432/ft_tran?schema=public" 
 
 JWT_SECRET="ishak-secret"
 BACKEND_PORT=5001
