+++ docker-compose.yml
@@ -19,7 +19,7 @@ services:
       context: ./backend
       target: development
     ports:
-      - "127.0.0.1:${BACKEND_PORT}:5001"
+      - "127.0.0.1:${BACKEND_PORT}:${BACKEND_PORT}"
       - "127.0.0.1:5555:5555" #for prisma Database Management
     depends_on:
       - postgres
@@ -35,7 +35,7 @@ services:
       context: ./Client
       target: development
     ports:
-      - "127.0.0.1:${FRONTEND_PORT}:3000"
+      - "127.0.0.1:${FRONTEND_PORT}:${FRONTEND_PORT}"
     depends_on:
       - backend
     networks:
